% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arms.R
\name{arms}
\alias{arms}
\title{Adaptive Rejection Metropolis Sampling}
\usage{
arms(n_samples, log_pdf, lower, upper, x_initial = lower + (1:n_initial) *
  (upper - lower)/(n_initial + 1), n_initial = 10, convex = 0,
  max_points = 100, metropolis = FALSE, x_previous = NA,
  include_neval = FALSE)
}
\arguments{
\item{n_samples}{Number of samples to return}

\item{log_pdf}{Potentially unnormalised log density from which to sample}

\item{lower}{Lower bound of the support of the distribution}

\item{upper}{Upper bound of the support of the distribution}

\item{x_initial}{Initial points with which to build the rejection distribution}

\item{n_initial}{Number of points used to form `x_initial`; ignored if `x_initial` provided.}

\item{convex}{Convexity adjustment}

\item{max_points}{Maximum number of points to allow in the rejection distribution}

\item{metropolis}{Whether to use a Metropolis-Hastings step after rejection sampling}

\item{x_previous}{The previous value of the Markov chain to be used if `metropolis = TRUE`}

\item{include_neval}{Whether to return an object specifying the number of function evaluations used}
}
\value{
Vector of samples if include_neval is FALSE, otherwise a list
}
\description{
Adaptive Rejection Metropolis Sampling
}
\references{
Gilks, W. R., Best, N. G. and Tan, K. K. C. (1995) Adaptive rejection Metropolis sampling. Applied Statistics, 44, 455-472.
}
\seealso{
\url{http://www1.maths.leeds.ac.uk/~wally.gilks/adaptive.rejection/web_page/Welcome.html}
}
